profile_service/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
pyjwt
flask-restx
prometheus_flask_exporter
marshmallow
flask_cors